requirements.txt
Pillow>=11.2.1
selenium>=4.32.0
webdriver-manager>=4.0.2